+++ src/Main/Aritmetica/Multiplicacion.java
@@ -12,7 +12,7 @@ package Main.Aritmetica;
 public class Multiplicacion implements Aritmetico{
 
     @Override
-    public float calcular(int a, int b) {
+    public int calcular(int a, int b) {
         return a*b;
     }
