+++ src/Main/Aritmetica/Resta.java
@@ -12,7 +12,7 @@ package Main.Aritmetica;
 public class Resta implements Aritmetico{
 
     @Override
-    public float calcular(int a, int b) {
+    public int calcular(int a, int b) {
         return a-b;
     }
